### Update `twitter:image` meta tag for correct property usage
Revised the `twitter:image` meta tag in the `index` file, replacing it with the correct property, `twitter:image:src`. This ensures compliance with Twitter's meta tag specifications and proper representation of branded visuals.

---

Updated the `twitter:image` meta tag to `twitter:image:src`, enhancing the integrity of the metadata and ensuring compatibility with Twitter's guidelines for social media previews.

---

#### Changes:

- **index**
  - Replaced `twitter:image` with `twitter:image:src` to adhere to Twitter's meta tag standards.

Signed-off-by: Zuzki <sleepy@zuzki.dev>