### Add image priority attribute in home hero image for improved loading
Updated the home hero image to include the `priority` attribute, enhancing performance by allowing faster loading of critical images on the page.

---

This small but important adjustment optimizes the loading behavior of the home page by instructing the browser to prioritize loading the hero image, potentially improving the user experience on initial load.

---

#### Changes:

- **home-hero.component.html**
  - Added the `priority` attribute to the hero image element for faster loading and increased performance focus.

Signed-off-by: Zuzki <sleepy@zuzki.dev>